debt - allow extHostSearch in common (#84109)
* debt - allow extHostSearch in common

* web - implement text search provider

* fix text search

* Update extensions/vscode-api-tests/src/extension.ts
