+++ src/vs/workbench/services/search/test/node/textSearchManager.test.ts
@@ -9,9 +9,9 @@ import { URI } from 'vs/base/common/uri';
 import { Progress } from 'vs/platform/progress/common/progress';
 import { ITextQuery, QueryType } from 'vs/workbench/services/search/common/search';
 import { ProviderResult, TextSearchComplete, TextSearchOptions, TextSearchProvider, TextSearchQuery, TextSearchResult } from 'vs/workbench/services/search/common/searchExtTypes';
-import { TextSearchManager } from 'vs/workbench/services/search/node/textSearchManager';
+import { NativeTextSearchManager } from 'vs/workbench/services/search/node/textSearchManager';
 
-suite('TextSearchManager', () => {
+suite('NativeTextSearchManager', () => {
 	test('fixes encoding', async () => {
 		let correctEncoding = false;
 		const provider: TextSearchProvider = {
@@ -33,7 +33,7 @@ suite('TextSearchManager', () => {
 			}]
 		};
 
-		const m = new TextSearchManager(query, provider);
+		const m = new NativeTextSearchManager(query, provider);
 		await m.search(() => { }, new CancellationTokenSource().token);
 
 		assert.ok(correctEncoding);
